// src/app/models/transfrac.ts
export interface transfrac {

    transactionId: number;
    dealerId: number;
    transDate: Date;
    nama: string;
    noPatG: string;
    type: string;
    weight: number;
    unitPrice: number;
    drc: number;
    total: number;
    totalOwner: number;
    totalTapper: number;
    noResit: string;
    companyAddress: string;
    tapperWithoutPatg: string;
    ownerPercentage: number;
}